refactor BoldedTitle component to a seperate folder
