+++ src/pages/tx/components/BoldedTitle.tsx
